consumer/requirements.txt
kafka-python
elasticsearch==7.10.1
